@@ -132,7 +132,9 @@ public class JSFImportGenerator {
 			b.append(EL_I18N_VAR + ".dialog_upload_title}\" widgetVar=\"" + widgetVar + "\">\n");
 			b.append("\t\t<p:messages id=\"" + msgVar + "\" showDetail=\"true\" showSummary=\"true\"/>\n");
 			b.append("\t\t<p:fileUpload auto=\"true\" mode=\"advanced\" listener=\"#{" + controller);
-			b.append("}\" sizeLimit=\"" + maxFileSize + "\" update=\"" + msgVar + "\"/>\n");
+			b.append("}\" update=\"" + msgVar + "\">\n");
+			b.append("\t\t\t<p:validateFile sizeLimit=\"" + maxFileSize + "\"/>\n");
+			b.append("\t\t</p:fileUpload>\n");
 			b.append("\t</p:dialog>\n\n");
 		}
 
