@@ -50,20 +50,18 @@ public class EclipseRAPPlugins {
 		final Set<PluginRecord> plugins = new HashSet<>();
 		plugins.add(new PluginRecord("com.ibm.icu"));
 		plugins.add(new PluginRecord("jakarta.annotation-api*2.1.1"));
-		plugins.add(new PluginRecord("jakarta.inject.jakarta.inject-api*2.0.1"));
+		plugins.add(new PluginRecord("jakarta.inject.jakarta.inject-api"));
 		plugins.add(new PluginRecord("jakarta.servlet-api"));
-		plugins.add(new PluginRecord("javax.servlet.jsp-api"));
 		plugins.add(new PluginRecord("org.apache.aries.spifly.dynamic.bundle"));
 		plugins.add(new PluginRecord("org.apache.commons.collections"));
 		plugins.add(new PluginRecord("org.apache.commons.commons-beanutils"));
-		plugins.add(new PluginRecord("org.apache.commons.commons-fileupload"));
+		plugins.add(new PluginRecord("org.apache.commons.commons-fileupload2-core"));
+		plugins.add(new PluginRecord("org.apache.commons.commons-fileupload2-jakarta-servlet6"));
 		plugins.add(new PluginRecord("org.apache.commons.commons-io"));
-		plugins.add(new PluginRecord("org.apache.commons.jxpath"));
 		plugins.add(new PluginRecord("org.apache.commons.logging"));
 		plugins.add(new PluginRecord("org.apache.felix.gogo.command"));
 		plugins.add(new PluginRecord("org.apache.felix.gogo.runtime"));
 		plugins.add(new PluginRecord("org.apache.felix.gogo.shell"));
-		plugins.add(new PluginRecord("org.apache.felix.http.servlet-api"));
 		plugins.add(new PluginRecord("org.apache.felix.scr", "1:true"));
 		plugins.add(new PluginRecord("org.eclipse.core.commands"));
 		plugins.add(new PluginRecord("org.eclipse.core.contenttype"));
@@ -98,33 +96,28 @@ public class EclipseRAPPlugins {
 		plugins.add(new PluginRecord("org.eclipse.equinox.common", "2:true"));
 		plugins.add(new PluginRecord("org.eclipse.equinox.console"));
 		plugins.add(new PluginRecord("org.eclipse.equinox.event"));
-		plugins.add(new PluginRecord("org.eclipse.equinox.http.jetty"));
-		plugins.add(new PluginRecord("org.eclipse.equinox.http.service.api"));
-		plugins.add(new PluginRecord("org.eclipse.equinox.http.servlet"));
 		plugins.add(new PluginRecord("org.eclipse.equinox.preferences"));
 		plugins.add(new PluginRecord("org.eclipse.equinox.registry"));
-		plugins.add(new PluginRecord("org.eclipse.jetty.ee8.security"));
-		plugins.add(new PluginRecord("org.eclipse.jetty.ee8.server"));
-		plugins.add(new PluginRecord("org.eclipse.jetty.ee8.servlet"));
+		plugins.add(new PluginRecord("org.eclipse.jetty.ee10.servlet"));
 		plugins.add(new PluginRecord("org.eclipse.jetty.http"));
 		plugins.add(new PluginRecord("org.eclipse.jetty.io"));
+		plugins.add(new PluginRecord("org.eclipse.jetty.logging"));
 		plugins.add(new PluginRecord("org.eclipse.jetty.security"));
 		plugins.add(new PluginRecord("org.eclipse.jetty.server"));
 		plugins.add(new PluginRecord("org.eclipse.jetty.session"));
 		plugins.add(new PluginRecord("org.eclipse.jetty.util.ajax"));
 		plugins.add(new PluginRecord("org.eclipse.jetty.util"));
-		plugins.add(new PluginRecord("org.eclipse.osgi.services"));
 		plugins.add(new PluginRecord("org.eclipse.osgi.util"));
 		plugins.add(new PluginRecord("org.eclipse.osgi", "1:true"));
 		plugins.add(new PluginRecord("org.eclipse.rap.e4"));
 		plugins.add(new PluginRecord("org.eclipse.rap.filedialog"));
 		plugins.add(new PluginRecord("org.eclipse.rap.fileupload"));
+		plugins.add(new PluginRecord("org.eclipse.rap.http.jetty"));
+		plugins.add(new PluginRecord("org.eclipse.rap.http.servlet"));
 		plugins.add(new PluginRecord("org.eclipse.rap.jface.databinding"));
 		plugins.add(new PluginRecord("org.eclipse.rap.jface"));
 		plugins.add(new PluginRecord("org.eclipse.rap.rwt.osgi"));
 		plugins.add(new PluginRecord("org.eclipse.rap.rwt"));
-		plugins.add(new PluginRecord("org.jdom"));
-		plugins.add(new PluginRecord("org.mortbay.jasper.apache-el"));
 		plugins.add(new PluginRecord("org.objectweb.asm.commons"));
 		plugins.add(new PluginRecord("org.objectweb.asm.tree.analysis"));
 		plugins.add(new PluginRecord("org.objectweb.asm.tree"));
@@ -134,7 +127,6 @@ public class EclipseRAPPlugins {
 		plugins.add(new PluginRecord("org.osgi.service.component"));
 		plugins.add(new PluginRecord("org.osgi.service.device"));
 		plugins.add(new PluginRecord("org.osgi.service.event"));
-		plugins.add(new PluginRecord("org.osgi.service.http.whiteboard"));
 		plugins.add(new PluginRecord("org.osgi.service.metatype"));
 		plugins.add(new PluginRecord("org.osgi.service.prefs"));
 		plugins.add(new PluginRecord("org.osgi.service.provisioning"));
@@ -147,7 +139,6 @@ public class EclipseRAPPlugins {
 		plugins.add(new PluginRecord("org.osgi.util.promise"));
 		plugins.add(new PluginRecord("org.osgi.util.xml"));
 		plugins.add(new PluginRecord("slf4j.api"));
-		plugins.add(new PluginRecord("slf4j.simple"));
 
 		return plugins;
 	}
