@@ -24,14 +24,10 @@ package net.codecadenza.runtime.webclient.vaadin.dialog;
 import static net.codecadenza.runtime.webclient.vaadin.i18n.InternalI18NService.CMD_OK;
 
 import com.vaadin.flow.component.AttachEvent;
-import com.vaadin.flow.component.Component;
 import com.vaadin.flow.component.button.Button;
+import com.vaadin.flow.component.button.ButtonVariant;
 import com.vaadin.flow.component.dialog.Dialog;
-import com.vaadin.flow.component.html.Div;
-import com.vaadin.flow.component.html.Span;
-import com.vaadin.flow.component.icon.Icon;
 import com.vaadin.flow.component.icon.VaadinIcon;
-import com.vaadin.flow.component.orderedlayout.HorizontalLayout;
 import java.util.Locale;
 import net.codecadenza.runtime.webclient.vaadin.i18n.InternalI18NService;
 
@@ -49,7 +45,6 @@ public abstract class AbstractTitleDialog extends Dialog {
 	public static final String DEFAULT_ICON_SIZE = "32px";
 	private static final long serialVersionUID = -6624572154912842951L;
 
-	private final Div divContent = new Div();
 	private final String title;
 	protected transient ButtonClickListener listener;
 	protected InternalI18NService i18n;
@@ -81,53 +76,25 @@ public abstract class AbstractTitleDialog extends Dialog {
 	 */
 	@Override
 	protected void onAttach(AttachEvent attachEvent) {
-		getElement().setAttribute("theme", "dialog-with-title-bar");
+		setHeaderTitle(title);
 
-		final var closeIcon = new Icon(VaadinIcon.CLOSE);
-		closeIcon.addClickListener(event -> close());
+		final Button closeButton = new Button(VaadinIcon.CLOSE.create(), _ -> close());
+		closeButton.addThemeVariants(ButtonVariant.LUMO_TERTIARY);
 
-		final var divTitleBar = new Div();
-		divTitleBar.addClassNames("titlebar-for-dialog", "draggable");
-		divTitleBar.add(new Span(title), closeIcon);
+		getHeader().add(closeButton);
 
-		divContent.addClassNames("content-for-dialog");
-
-		final var divButtonBar = new Div();
-		divButtonBar.addClassNames("buttonbar-for-dialog");
-
-		final var hlButtons = new HorizontalLayout();
-		hlButtons.setPadding(true);
-
-		addButtons(hlButtons);
-
-		super.add(divTitleBar, divContent, divButtonBar, hlButtons);
+		addButtons(getFooter());
 	}
 
 	/**
 	 * @param hlButtons
 	 */
-	protected void addButtons(HorizontalLayout hlButtons) {
+	protected void addButtons(DialogFooter footer) {
 		final var cmdOK = new Button(i18n.getTranslation(CMD_OK));
 		cmdOK.setId("cmdOK");
-		cmdOK.addClickListener(event -> fireButtonClickEvent(ButtonType.OK));
+		cmdOK.addClickListener(_ -> fireButtonClickEvent(ButtonType.OK));
 
-		hlButtons.add(cmdOK);
-	}
-
-	/*
-	 * (non-Javadoc)
-	 * @see com.vaadin.flow.component.dialog.Dialog#add(com.vaadin.flow.component.Component[])
-	 */
-	@Override
-	public void add(Component... components) {
-		divContent.add(components);
-	}
-
-	/**
-	 * @return the component where the specific dialog content should be added to
-	 */
-	public Div getContentArea() {
-		return divContent;
+		footer.add(cmdOK);
 	}
 
 	/**
