@@ -28,7 +28,6 @@ import static net.codecadenza.runtime.webclient.vaadin.i18n.InternalI18NService.
 import com.vaadin.flow.component.button.Button;
 import com.vaadin.flow.component.icon.Icon;
 import com.vaadin.flow.component.icon.VaadinIcon;
-import com.vaadin.flow.component.orderedlayout.HorizontalLayout;
 import java.util.Locale;
 
 /**
@@ -65,23 +64,24 @@ public class ConfirmationMessageDialog extends AbstractMessageDialog {
 
 	/*
 	 * (non-Javadoc)
-	 * @see net.codecadenza.runtime.webclient.vaadin.dialog.AbstractMessageDialog#addButtons(com.vaadin.ui.HorizontalLayout)
+	 * @see net.codecadenza.runtime.webclient.vaadin.dialog.AbstractTitleDialog#
+	 * addButtons(com.vaadin.flow.component.dialog.Dialog.DialogFooter)
 	 */
 	@Override
-	protected void addButtons(HorizontalLayout buttonLayout) {
+	protected void addButtons(DialogFooter dialogFooter) {
 		final var cmdYes = new Button(i18n.getTranslation(CMD_YES));
 		cmdYes.setId("cmdYes");
-		cmdYes.addClickListener(event -> fireButtonClickEvent(ButtonType.YES));
+		cmdYes.addClickListener(_ -> fireButtonClickEvent(ButtonType.YES));
 
 		final var cmdNo = new Button(i18n.getTranslation(CMD_NO));
 		cmdNo.setId("cmdNo");
-		cmdNo.addClickListener(event -> fireButtonClickEvent(ButtonType.NO));
+		cmdNo.addClickListener(_ -> fireButtonClickEvent(ButtonType.NO));
 
 		final var cmdCancel = new Button(i18n.getTranslation(CMD_CANCEL));
 		cmdCancel.setId("cmdCancel");
-		cmdCancel.addClickListener(event -> fireButtonClickEvent(ButtonType.CANCEL));
+		cmdCancel.addClickListener(_ -> fireButtonClickEvent(ButtonType.CANCEL));
 
-		buttonLayout.add(cmdYes, cmdNo, cmdCancel);
+		dialogFooter.add(cmdYes, cmdNo, cmdCancel);
 	}
 
 }
